Crate spin_factor_outbound_networking

Re-exports

pub use runtime_config::ComponentTlsConfigs;

Modules

runtime_config

Structs

AllowedHostConfig

An address is a url-like string that contains a host, a port, and an optional scheme

AppState

InstanceBuilder

OutboundAllowedHosts

A check for whether a URL is allowed by the outbound networking configuration.

OutboundNetworkingFactor

OutboundUrl

A parsed URL used for outbound networking.

Enums

AllowedHostsConfig

HostConfig

Constants

SERVICE_CHAINING_DOMAIN_SUFFIX

Traits

DisallowedHostHandler

Functions

allowed_outbound_hosts

Get the raw values of the allowed_outbound_hosts locked app metadata key.

is_service_chaining_host

parse_service_chaining_target

record_address_fields

Records the address host, port, and database as fields on the current tracing span.

validate_service_chaining_for_components

Validates that all service chaining of an app will be satisfied by the supplied subset of components.

Type Aliases

SharedFutureResult